Add true realtime backoffice queue updates
