+++ docs/BACKOFFICE.md
@@ -210,11 +210,13 @@ intentionally conservative:
   `catalog-maintenance` queue, including waiting, active, scheduled, failed, and
   completed jobs with compact payload fields;
 - the queue page listens to BullMQ `QueueEvents` through a server-sent events
-  stream that pushes the current queue snapshot for the active filter/page, so
-  waiting, active, completed, failed, delayed, and stalled job changes update
-  the operator view without waiting for manual refresh;
+  stream and applies the current queue snapshot for the active filter/page
+  directly in the browser, so waiting, active, completed, failed, delayed, and
+  stalled job changes update the operator view without waiting for manual
+  refresh or a full page reload;
 - the catalog-health home uses a dedicated server-sent events stream that pushes
-  the live DB and queue snapshot after catalog-maintenance changes, with a
+  the live DB and queue snapshot after catalog-maintenance changes. Its status,
+  queue counts, and summary cards update directly from the live snapshot, with a
   slower background check only as a reconnect fallback;
 - the repair batch history page at `/repair-batches` shows recent durable batch
   attempts and links to per-item details, so operators do not need to infer
